Move horizontal ColorBar stories into separate file
